+++ lib/kandr-rails-extensions/string.rb
@@ -7,6 +7,14 @@ module KandrRailsExtensions
     end
   end
 
+  module StringToBoolean
+    def to_bool
+      return true if self == true || self =~ (/^(true|t|yes|y|on|1)$/i)
+      return false if self == false || self.blank? || self =~ (/^(false|f|no|n|off|0)$/i)
+      raise ArgumentError.new("invalid value for Boolean: \"#{self}\"")
+    end
+  end
+
   module StringSlugify
     def slugify
       # strip the string
@@ -61,5 +69,6 @@ end
 class String
   include KandrRailsExtensions::StringSlugify
   include KandrRailsExtensions::StringTitleize
+  include KandrRailsExtensions::StringToBoolean
   include KandrRailsExtensions::StringTrim
 end